chore: move unit methods to new file on application domain
The application.go file in the service layer of application was getting
way too big so this patch moves the unit-related methods to a new
unit.go file.
Tests are also moved to unit_test.go.
